// cpo-basic-app-front/src/apis/usermanage/pool/poolUrl.ts
enum poolUrl {
  list = '/sys/user/page',
  add = '/sys/user/insert',
  update = '/sys/user/update',
  delete = '/sys/user/delete',
  reset = '/sys/user/updatePassword',
  change = '/sys/user/changeActive',
  check = '/sys/user/checkPassword',
  getDefault = '/sys/user/getDefaultPwd',
  resetpasswd = '/sys/user/resetPassword'
}

export default poolUrl